Rock_Paper_Scissors_Game
A simple "Rock, Paper, Scissors" game. This program takes the input of the user and compares it against the input that is randomly attributed to the variable "computer". It utilizes while loops, if, elif and nested if statements for this purpose. In order for the game not go on forever, I implemented a break from the while loop when either the player or the computer achieved 3 victories. For that purpose there are also variables that store the number of victories for each of the "competitors" and increment after each victory.