Arithmetic Calculator
This project was built with the goal of providing the user with a calculator which has an easy to use interface. It provides the user with a basic way to calculate arithmetic operations between two numbers. This program will also record each operation into a text file and allow the user to request to read it.